+++ src/main/java/com/mobble/mobbleserver/application/comment/service/CommentModifyService.java
@@ -11,7 +11,7 @@ import com.mobble.mobbleserver.application.comment.port.provided.CommentDeletePo
 import com.mobble.mobbleserver.application.comment.port.provided.CommentUpdatePort;
 import com.mobble.mobbleserver.application.comment.port.required.CommentReadPort;
 import com.mobble.mobbleserver.application.comment.port.required.CommentWritePort;
-import com.mobble.mobbleserver.application.exception.BusinessException;
+import com.mobble.mobbleserver.application.common.exception.BusinessException;
 import com.mobble.mobbleserver.domain.article.Article;
 import com.mobble.mobbleserver.domain.clubMember.ClubMember;
 import com.mobble.mobbleserver.domain.comment.Comment;
